docs: rewrite sd-step-group documentation
